refactor to remove magic numbers
