+++ core/src/main/java/io/questdb/cairo/TxReader.java
@@ -109,7 +109,7 @@ public class TxReader implements Closeable, Mutable {
         clear();
     }
 
-    public void dumpPartitionInfo(LongList container) {
+    public void dumpRawTxPartitionInfo(LongList container) {
         container.add(attachedPartitions);
     }
 
@@ -278,7 +278,7 @@ public class TxReader implements Closeable, Mutable {
     }
 
     public long getPartitionNameTxnByRawIndex(int index) {
-        return attachedPartitions.getQuick(index + PARTITION_NAME_TX_OFFSET);
+        return getPartitionNameTxnByRawIndex(attachedPartitions, index);
     }
 
     public long getPartitionParquetFileSize(int partitionIndex) {
@@ -298,7 +298,7 @@ public class TxReader implements Closeable, Mutable {
     }
 
     public long getPartitionSizeByRawIndex(int index) {
-        return attachedPartitions.getQuick(index + PARTITION_MASKED_SIZE_OFFSET) & PARTITION_SIZE_MASK;
+        return getPartitionSizeByRawIndex(attachedPartitions, index);
     }
 
     public long getPartitionTableVersion() {
@@ -368,6 +368,10 @@ public class TxReader implements Closeable, Mutable {
         return isPartitionParquetByRawIndex(i * LONGS_PER_TX_ATTACHED_PARTITION);
     }
 
+    public boolean isPartitionParquetByRawIndex(int indexRaw) {
+        return checkPartitionOptionBit(indexRaw, PARTITION_MASK_PARQUET_FORMAT_BIT_OFFSET);
+    }
+
     public boolean isPartitionReadOnly(int i) {
         return isPartitionReadOnlyByRawIndex(i * LONGS_PER_TX_ATTACHED_PARTITION);
     }
@@ -378,6 +382,10 @@ public class TxReader implements Closeable, Mutable {
             return isPartitionReadOnlyByRawIndex(indexRaw);
         }
         return false;
+    }
+
+    public boolean isPartitionReadOnlyByRawIndex(int indexRaw) {
+        return checkPartitionOptionBit(indexRaw, PARTITION_MASK_READ_ONLY_BIT_OFFSET);
     }
 
     public TxReader ofRO(@Transient LPSZ path, int partitionBy) {
@@ -530,14 +538,6 @@ public class TxReader implements Closeable, Mutable {
         return attachedPartitions.getQuick(partitionRawIndex + PARTITION_PARQUET_FILE_SIZE_OFFSET);
     }
 
-    public boolean isPartitionParquetByRawIndex(int indexRaw) {
-        return checkPartitionOptionBit(indexRaw, PARTITION_MASK_PARQUET_FORMAT_BIT_OFFSET);
-    }
-
-    public boolean isPartitionReadOnlyByRawIndex(int indexRaw) {
-        return checkPartitionOptionBit(indexRaw, PARTITION_MASK_READ_ONLY_BIT_OFFSET);
-    }
-
     private void openTxnFile(FilesFacade ff, LPSZ path) {
         if (ff.exists(path)) {
             if (roTxMemBase == null) {
@@ -596,6 +596,18 @@ public class TxReader implements Closeable, Mutable {
         for (int i = 0; i < symbolMapCount; i++) {
             symbolCountSnapshot.add(getInt(TableUtils.getSymbolWriterIndexOffset(i)));
         }
+    }
+
+    static int findPartitionRawIndex(LongList attachedPartitions, long partitionTimestamp) {
+        return attachedPartitions.binarySearchBlock(LONGS_PER_TX_ATTACHED_PARTITION_MSB, partitionTimestamp, BinarySearch.SCAN_UP);
+    }
+
+    static long getPartitionNameTxnByRawIndex(LongList attachedPartitions, int index) {
+        return attachedPartitions.getQuick(index + PARTITION_NAME_TX_OFFSET);
+    }
+
+    static long getPartitionSizeByRawIndex(LongList attachedPartitions, int index) {
+        return attachedPartitions.getQuick(index + PARTITION_MASKED_SIZE_OFFSET) & PARTITION_SIZE_MASK;
     }
 
     protected int findAttachedPartitionRawIndex(long ts) {
